(editor) Extract shared form-control renderers
Move the native input/button renderers (_renderNumberField, _renderTextField,
_renderResetButton, _renderTextButton) into src/editor/field-renderers.js as
free functions; base keeps thin delegations. Verbatim move.
